requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pymilvus==2.3.0
sentence-transformers==2.2.2
huggingface-hub<0.19.0
python-dotenv==1.0.0
pydantic==2.3.0
numpy==1.25.2
PyPDF2==3.0.1
marshmallow==3.17.0
jinja2==3.1.2
requests==2.31.0
loguru==0.7.0
